requirements.txt
msal>=1.30.0
requests>=2.32.0
openai>=1.35.0
pdfminer.six>=20231228
python-docx>=1.1.2
